==> notebooks/0-Data_Exploration.ipynb <==
# Data description

---


This project aimed at providing a data-driven solution to enable students to gain insights into the relationships between different functions and their associated technologies.

**Data Source**:

- The project will leverage the Stack Overflow Developer Survey 2023 dataset, as it provides a rich source of information regarding the skills and technologies associated with various IT jobs. This dataset will serve as the foundation for the data-driven analysis and recommendations provided by SkillMap.

- Data link: https://www.kaggle.com/datasets/stackoverflow/stack-overflow-2023-developers-survey?select=survey_results_public.csv


## Data schema


1. **ResponseId (or equivalent identifier):** Unique identifier for each survey response.

2. **MainBranch:** Describes the respondent's current role or branch (e.g., developer by profession).

3. **Age:** Provides information about the respondent's age.

4. **Employment:** Describes the respondent's current employment status.

5. **RemoteWork:** Indicates the respondent's current work situation regarding remote work.

6. **CodingActivities:** Describes the code-related activities the respondent engages in outside of work.

7. **EdLevel:** Specifies the highest level of formal education completed by the respondent.

8. **LearnCode, LearnCodeOnline, LearnCodeCoursesCert:** Describes how the respondent learns to code, including online resources and courses.

9. **YearsCode, YearsCodePro:** Indicates the total years of coding experience, including professional and non-professional.

10. **DevType:** Describes the respondent's current job role or function.

11. **OrgSize:** Specifies the size of the organization where the respondent works.

12. **PurchaseInfluence:** Indicates the level of influence the respondent has over new technology purchases in their organization.

13. **TechList:** Describes the respondent's approach when thinking about new technology purchases.

14. **BuyNewTool:** Describes how the respondent discovers and researches new tools or software.

15. **Country:** Specifies the respondent's location or country.

16. **Currency:** Indicates the currency used day-to-day by the respondent.

17. **CompTotal:** Provides information about the respondent's current total annual compensation.

18. **LanguageHaveWorkedWith, LanguageWantToWorkWith:** Lists programming languages the respondent has worked with and wants to work with.

19. **DatabaseHaveWorkedWith, DatabaseWantToWorkWith:** Lists databases the respondent has worked with and wants to work with.

20. **PlatformHaveWorkedWith, PlatformWantToWorkWith:** Lists platforms the respondent has worked with and wants to work with.

21. **WebframeHaveWorkedWith, WebframeWantToWorkWith:** Lists web frameworks the respondent has worked with and wants to work with.

22. **MiscTechHaveWorkedWith, MiscTechWantToWorkWith:** Lists miscellaneous technologies the respondent has worked with and wants to work with.

23. **ToolsTechHaveWorkedWith, ToolsTechWantToWorkWith:** Lists tools and technologies the respondent has worked with and wants to work with.

24. **NEWCollabToolsHaveWorkedWith, NEWCollabToolsWantToWorkWith:** Lists collaborative tools the respondent has worked with and wants to work with.

25. **OfficeStackAsyncHaveWorkedWith, OfficeStackAsyncWantToWorkWith:** Lists asynchronous office stack tools the respondent has worked with and wants to work with.

26. **OfficeStackSyncHaveWorkedWith, OfficeStackSyncWantToWorkWith:** Lists synchronous office stack tools the respondent has worked with and wants to work with.

27. **AISearchHaveWorkedWith, AISearchWantToWorkWith:** Lists AI-powered search tools the respondent has worked with and wants to work with.

28. **AIDevHaveWorkedWith, AIDevWantToWorkWith:** Lists AI-powered developer tools the respondent has worked with and wants to work with.

29. **NEWSOSites:** Lists Stack Overflow sites visited by the respondent.

30. **SOVisitFreq, SOAccount, SOPartFreq, SOComm:** Describes the respondent's engagement with Stack Overflow.

31. **ConvertedCompYearly:** Indicates the respondent's yearly compensation.

32. **SurveyLength, SurveyEase:** Describes the respondent's perception of the survey length and ease.

33. **Knowledge_1 to Knowledge_8:** Ratings indicating the respondent's level of agreement with various statements related to knowledge and workflow.

34. **Frequency_1 to Frequency_3:** Ratings indicating the frequency of certain experiences at work.

35. **TimeSearching, TimeAnswering:** Indicate the average time spent searching for answers and answering questions at work.

36. **ProfessionalTech, Industry, SOTeamsUsage:** Provide information about the respondent's company and industry.

37. **TBranch, ICorPM, WorkExp:** Describe the respondent's willingness to participate in the Professional Developer Series, individual contributor or people manager status, and years of working experience.

38. **S7 (Final Questions), SurveyLength, SurveyEase:** Provide additional insights into the respondent's perception of the survey.


# Problem Description


- In the fast-evolving landscape of IT jobs and technologies, students often find themselves bewildered by the multitude of skills associated with different roles. Questions like "Do I need to learn C++ to be a Data Scientist?" or "Can I use JavaScript in Data Analytics?" reflect the confusion that aspiring professionals face.

- Our client is an IT educational institute that asked us to create a project aimed at providing a data-driven solution to enable students to gain insights into the relationships between different functions and their associated technologies.


# Data undestinding

---


DATA_PATH = r'..\data\raw\survey_results_public.csv'

## Importing


import pandas as pd
import numpy as np

# Set the maximum number of rows and columns to display
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)

df = pd.read_csv(DATA_PATH)

df.head()

df.info()

df.iloc[4]

df.describe().T

## Well, at first glance, the data contains a lot of columns that are not useful for our case. Therefore, you should select the features that are most relevant to our problem first.


## will do that in next notebook --> feature_selection.ipynb
