==> notebooks/1_1-Features_selection.ipynb <==
# Features selection


Our project aims to ensure that if a student enters their skills, our model will respond with the job title most relevant to their skills, so we will focus on features that include that information


DATA_PATH = r'..\data\raw\survey_results_public.csv'
OUTPUT_DATA_PATH = r'..\data\processed\01-relevant_data.csv'

## Importing


import pandas as pd
import numpy as np

# Set the maximum number of rows and columns to display
# pd.set_option('display.max_rows', None)
# pd.set_option('display.max_columns', None)

df = pd.read_csv(DATA_PATH)

df.head()

relevant_columns = [
    'LanguageHaveWorkedWith', 'LanguageWantToWorkWith',
    'DatabaseHaveWorkedWith', 'DatabaseWantToWorkWith',
    'PlatformHaveWorkedWith', 'PlatformWantToWorkWith',
    'WebframeHaveWorkedWith', 'WebframeWantToWorkWith',
    'MiscTechHaveWorkedWith', 'MiscTechWantToWorkWith',
    'Country',
    'EdLevel',
    'YearsCode', 'YearsCodePro',
    'DevType'
]

relevant_columns = df[relevant_columns]

relevant_columns.shape

relevant_columns.iloc[1]

relevant_columns

relevant_columns.to_csv(OUTPUT_DATA_PATH)

Here's a description for each column in the `relevant_columns` list:

1. `LanguageHaveWorkedWith`: The programming, scripting, and markup languages that the survey respondent has worked with.

2. `LanguageWantToWorkWith`: The programming, scripting, and markup languages that the survey respondent wants to work with in the future.

3. `DatabaseHaveWorkedWith`: The database environments that the survey respondent has worked with.

4. `DatabaseWantToWorkWith`: The database environments that the survey respondent wants to work with in the future.

5. `PlatformHaveWorkedWith`: The cloud platforms that the survey respondent has worked with.

6. `PlatformWantToWorkWith`: The cloud platforms that the survey respondent wants to work with in the future.

7. `WebframeHaveWorkedWith`: The web frameworks and web technologies that the survey respondent has worked with.

8. `WebframeWantToWorkWith`: The web frameworks and web technologies that the survey respondent wants to work with in the future.

9. `MiscTechHaveWorkedWith`: Other frameworks and libraries that the survey respondent has worked with.

10. `MiscTechWantToWorkWith`: Other frameworks and libraries that the survey respondent wants to work with in the future.

11. `Country`: The country where the survey respondent is located.

12. `EdLevel`: The highest level of formal education that the survey respondent has completed.

13. `YearsCode`: The total number of years the survey respondent has been coding, including any education.

14. `YearsCodePro`: The number of years the survey respondent has been coding professionally, excluding education.

15. `DevType`: The current job description or role of the survey respondent, representing the type of developer they are.
