==> notebooks/2-Feature Engineering.ipynb <==
# Feature Engineering


DATA_PATH = r'..\data\processed\02-Cleaned_data.csv'
OUTPUT_DATA_PATH = r'..\data\processed\03-final_data.csv'
OUTPUT_DATA_PATH_pickle = r'..\data\processed\03-final_data.pkl'

# Functions

## Importing


import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pylab as plt
# Set the maximum number of rows and columns to display
# pd.set_option('display.max_rows', None)
# pd.set_option('display.max_columns', None)

df = pd.read_csv(DATA_PATH)

df.head()

df.iloc[0]

There are multiple answers in the some row so we want to convert it to list


## -Convert multiple answers to list


text = 'HTML/CSS;JavaScript;Python'

text.split(';')

columns_to_split = ['LanguageHaveWorkedWith', 'LanguageWantToWorkWith',
                    'DatabaseHaveWorkedWith', 'PlatformHaveWorkedWith', 'WebframeHaveWorkedWith']

# Apply str.split(';') to each specified column
for column in columns_to_split:
    df[column] = df[column].str.split(';')

df.head()

type(df['LanguageHaveWorkedWith'].iloc[0])

df.info()

## -Merge LanguageHaveWorkedWith with LanguageWantToWorkWith and remove duplicate skills


df[df['DevType'] == 'Data scientist or machine learning specialist'].sample().iloc[0]

x = ['HTML/CSS', 'Python', 'Swift', 'TypeScript']
y = ['HTML/CSS', 'Python', 'TypeScript']

list(set(x + y))

df.apply(lambda row: list(
    set(row['LanguageHaveWorkedWith'] + row['LanguageWantToWorkWith'])), axis=1)

# Assuming df is your DataFrame
df['Key_languages'] = df.apply(lambda row: list(
    set(row['LanguageHaveWorkedWith'] + row['LanguageWantToWorkWith'])), axis=1)

df.head()

df[df['DevType'] == 'Data scientist or machine learning specialist'].sample().iloc[0]

df.drop(['LanguageHaveWorkedWith', 'LanguageWantToWorkWith'], axis=1, inplace=True)

## -Rename columns


df.head()

df = df.rename(columns={'DatabaseHaveWorkedWith': 'Databases', 'PlatformHaveWorkedWith': 'Platforms', 'WebframeHaveWorkedWith': 'Web_frames',
                        'YearsCode': 'Total_years', 'YearsCodePro': 'experanse_years'})

df.columns

df = df[['Key_languages', 'Databases', 'Platforms', 'Web_frames', 'Country', 'EdLevel',
         'experanse_years', 'Total_years', 'DevType']]

df.head()

## -map the values in the "EdLevel" column to simpler names


df['EdLevel'].value_counts()

# Mapping dictionary
ed_level_mapping = {
    'Bachelor’s degree (B.A., B.S., B.Eng., etc.)': 'Bachelor\'s',
    'Master’s degree (M.A., M.S., M.Eng., MBA, etc.)': 'Master\'s',
    'Some college/university study without earning a degree': 'Some College/University',
    'Professional degree (JD, MD, Ph.D, Ed.D, etc.)': 'Professional',
    'Secondary school (e.g. American high school, German Realschule or Gymnasium, etc.)': 'Secondary School',
    'Associate degree (A.A., A.S., etc.)': 'Associate',
    'Something else': 'Other',
    'Primary/elementary school': 'Primary/Elementary'
}

# Map the values in the 'EdLevel' column
df['EdLevel'] = df['EdLevel'].map(ed_level_mapping)

df.head()

## - map the values in the 'DevType' column to simpler names


df['DevType'].unique()

# Sample mapping dictionary
devtype_mapping = {
    'Senior Executive (C-Suite, VP, etc.)': 'Executive',
    'Developer, back-end': 'Back-end Developer',
    'Developer, front-end': 'Front-end Developer',
    'Developer, full-stack': 'Full-stack Developer',
    'System administrator': 'System Administrator',
    'Developer, desktop or enterprise applications': 'Desktop/Enterprise Developer',
    'Developer, QA or test': 'QA/Test Developer',
    'Designer': 'Designer',
    'Data scientist or machine learning specialist': 'Data Scientist/ML Specialist',
    'Data or business analyst': 'Data/Business Analyst',
    'Security professional': 'Security Professional',
    'Research & Development role': 'R&D Developer',
    'Other (please specify):': 'Other',
    'Developer, mobile': 'Mobile Developer',
    'Database administrator': 'Database Administrator',
    'Developer, embedded applications or devices': 'Embedded Developer',
    'Engineer, data': 'Data Engineer',
    'Hardware Engineer': 'Hardware Engineer',
    'Product manager': 'Product Manager',
    'Academic researcher': 'Academic Researcher',
    'Developer, game or graphics': 'Game/Graphics Developer',
    'Cloud infrastructure engineer': 'Cloud Infrastructure Engineer',
    'Student': 'Student',
    'Developer Experience': 'Developer Experience',
    'Engineering manager': 'Engineering Manager',
    'Project manager': 'Project Manager',
    'DevOps specialist': 'DevOps Specialist',
    'Engineer, site reliability': 'Site Reliability Engineer',
    'Blockchain': 'Blockchain Developer',
    'Developer Advocate': 'Developer Advocate',
    'Educator': 'Educator',
    'Scientist': 'Scientist',
    'Marketing or sales professional': 'Marketing/Sales Professional',
}

# Apply the mapping to the 'DevType' column
df['DevType'] = df['DevType'].map(devtype_mapping)

# df = df[~(df['DevType'] =='Other')]

df[(df['DevType'] == 'Other')]

df.head()

## -Make new column (learning years)


df['learning_years'] = df['Total_years'] - df['experanse_years']

df.head()

df = df[['Key_languages', 'Databases', 'Platforms', 'Web_frames', 'Country', 'EdLevel',
         'experanse_years', 'learning_years', 'Total_years', 'DevType']]

df

# The data is ready to analysis now


---


df.to_csv(OUTPUT_DATA_PATH, index=False)

df.to_pickle(OUTPUT_DATA_PATH_pickle)

we save as pickle file to keep all data type the same
