==> notebooks/4_2-Data preprocessing.ipynb <==
# Data preprocessing

---


DATA_PATH = r'..\data\processed\04-final_data_v2.pkl'


DATA_X_TRAIN = r'..\data\processed\X_train.pkl'

DATA_X_TEST = r'..\data\processed\X_test.pkl'

DATA_y_TRAIN = r'..\data\processed\y_train.pkl'

DATA_y_TEST = r'..\data\processed\y_test.pkl'



list_columns = ['Key_languages', 'Databases', 'Web_frames', 'Platforms']

specified_columns = ['EdLevel', 'experanse_years']

## Importing


import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pylab as plt

from imblearn.over_sampling import RandomOverSampler
from imblearn.under_sampling import RandomUnderSampler

from sklearn.pipeline import Pipeline

from sklearn.preprocessing import StandardScaler, MinMaxScaler, RobustScaler
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.compose import ColumnTransformer

# Set the maximum number of rows and columns to display
# pd.set_option('display.max_rows', None)
# pd.set_option('display.max_columns', None)

# Functions
def encoding_list_col(df):
    # Perform one-hot encoding on list columns

    for column in list_columns:
        # Create a list of unique values from all lists in the column
        unique_values = set(
            value for sublist in df[column] for value in sublist)

        # Create binary features for each unique value
        for value in unique_values:
            df[f"{column}_{value}"] = df[column].apply(
                lambda x: int(value in x))

    # Drop the original list columns
    df = df.drop(columns=list_columns)

    # Rearrange columns to move specified columns to the end

    df = pd.concat([df.drop(columns=specified_columns),
                    df[specified_columns]], axis=1)

    return df

df = pd.read_pickle(DATA_PATH)

### 1)Feature selection and Engineering


df.head()

df = encoding_list_col(df)

df['DevType'].value_counts()

df = df[~((df['DevType'] == 'Marketing/Sales Professional') | (df['DevType'] == 'Developer Advocate') | (df['DevType'] == 'Designer') | (df['DevType'] == 'Scientist')
          | (df['DevType'] == 'Educator') | (df['DevType'] == 'Product Manager') | (df['DevType'] == 'Project Manager') | (df['DevType'] == 'Developer Experience ') | (df['DevType'] == 'Engineering Manager')
          )]

df['DevType'].value_counts()

Done


### 2)split X and y


X = df.drop('DevType', axis=1)
y = df['DevType']

### **Balance classes:Resample**


samples_per_class = 1400
y_cont = y.value_counts()

for index, values in y_cont.iteritems():
    if values > samples_per_class:
        under_s = RandomUnderSampler(
            sampling_strategy={index: 1713}, random_state=42)
        X, y = under_s.fit_resample(X, y)

    elif values < samples_per_class:
        under_s = RandomOverSampler(
            sampling_strategy={index: 1377}, random_state=42)
        X, y = under_s.fit_resample(X, y)

y.value_counts()

### 3)Trian test split


from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.15, random_state=42)

y_train.value_counts()

### 4)preprocessing pipeline


num_feat = X_train.select_dtypes(exclude='object').columns
cat_feat = X_train.select_dtypes(include='object').columns

scaler = RobustScaler()
encoder = OneHotEncoder()


num_pipeline = Pipeline(
    steps=[
        ('scaler', scaler)
    ]
)

cat_pipeline = Pipeline(
    steps=[
        ('one_hot_encoder', OneHotEncoder(drop='first'))
        # Set with_mean=False for sparse input
        # ('scaler', StandardScaler(with_mean=False)),
    ]
)

preperocessor = ColumnTransformer(
    [
        ('num_pipeline', num_pipeline, num_feat),
        ('cat_pipeline', cat_pipeline, cat_feat)
    ]
)

X_train_arr = preperocessor.fit_transform(X_train)
X_test_arr = preperocessor.transform(X_test)

X_train_arr.shape

X_train_df = pd.DataFrame(
    X_train_arr, columns=preperocessor.get_feature_names_out())
X_test_df = pd.DataFrame(
    X_test_arr, columns=preperocessor.get_feature_names_out())

# X_train_df

---


### Save data


X_train_df.to_pickle(DATA_X_TRAIN)
X_test_df.to_pickle(DATA_X_TEST)

# pd.to_pickle(X_train_arr , DATA_X_TRAIN)
# pd.to_pickle(X_test_arr , DATA_X_TEST)

# pd.to_pickle(y_train , y_train)
# pd.to_pickle(y_test , DATA_y_TEST)


y_train.to_pickle(DATA_y_TRAIN)
y_test.to_pickle(DATA_y_TEST)

**Conclusion:**
The notebook successfully addresses several crucial steps in the data preprocessing phase, including feature engineering, class balancing, and creating an efficient preprocessing pipeline. It also demonstrates good practices such as stratified sampling and saving intermediate data.
