==> notebooks/5_1-modeling-basemodel.ipynb <==
DATA_PATH = r'..\data\processed\04-final_data_v2.pkl'
DATA_X_TRAIN = r'..\data\processed\X_train.pkl'
DATA_X_TEST = r'..\data\processed\X_test.pkl'
DATA_y_TRAIN = r'..\data\processed\y_train.pkl'
DATA_y_TEST = r'..\data\processed\y_test.pkl'

## Importing


import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pylab as plt

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.decomposition import PCA
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier
from sklearn.svm import SVC

# Set the maximum number of rows and columns to display
# pd.set_option('display.max_rows', None)
# pd.set_option('display.max_columns', None)

X_train = pd.read_pickle(DATA_X_TRAIN)
X_test = pd.read_pickle(DATA_X_TEST)

y_train = pd.read_pickle(DATA_y_TRAIN)
y_test = pd.read_pickle(DATA_y_TEST)

X_train

# Baseline model

The baseline model phase is a crucial step in machine learning projects where you establish a basic model as a starting point for comparison.

---


### 1)choose model


we will choose logistic regression model as simplest model


pca_model = PCA(n_components=90)
model = LogisticRegression(max_iter=1000)

### 2)Train the model


X_train_pca = pca_model.fit_transform(X_train)
X_test_pca = pca_model.transform(X_test)

len(X_train_pca)

pca_model.explained_variance_ratio_.sum()

y_train.value_counts()

model.fit(X_train_pca, y_train)

pred = model.predict(X_test_pca)

pred_train = model.predict(X_train_pca)

### 3)Evaluate the model


from sklearn.metrics import classification_report, confusion_matrix, accuracy_score

accuracy_score(y_train, pred_train)

accuracy_score(y_test, pred)

print(classification_report(y_test, pred))

The baseline model don't perform well, so let's move on to choosing model notebook
