==> notebooks/6_1-prediction.ipynb <==
DATA_PATH = r'..\data\processed\04-final_data_v2.pkl'


X_data = r'..\data\processed\Train And test F_selection\X_Data.pkl'
y_data = r'..\data\processed\Train And test F_selection\y_Data.pkl'

preperocessor_path = r'..\models\preperocessor.pkl'
final_model_path = r'..\models\final_model.pkl'


DATA_COL = r'..\models\data_col.pkl'
Classes_NAME = r'..\models\classes_names.pkl'


LANGUAGES_PATH = r'..\models\data details\languages.pkl'
DATABASES_PATH = r'..\models\data details\DATABASES.pkl'
PLATFROMS_PATH = r'..\models\data details\PLATFROMS.pkl'
WEB_FRAMES_PATH = r'..\models\data details\WEB_FRAMES.pkl'
VARIED_TACH_PATH = r'..\models\data details\VARIED_TACH.pkl'
TOOLS_PATH = r'..\models\data details\TOOLS.pkl'
COLLABTOOLS_PATH = r'..\models\data details\COLLABTOOLS.pkl'


skills_clusters_path = r'..\models\data details\skills_clusters.pkl'


list_columns = ['Key_languages', 'Databases', 'Platforms',
                'Web_frames', 'varied_Tech', 'Tools', 'collabTools']

specified_columns = ['experanse_years']

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pylab as plt
from sklearn.preprocessing import StandardScaler


import pickle
import joblib

# Set the maximum number of rows and columns to display
pd.set_option('display.max_rows', 500)
pd.set_option('display.max_columns', 500)

df = pd.read_pickle(DATA_PATH)
X = pd.read_pickle(X_data)
y = pd.read_pickle(y_data)

skills_list = ['R', 'Python', 'TensorFlow',
               'Jupyter Notebook/JupyterLab', 'pandas', 'Numpy']
experanse_years = 0

# Make prediction fun


def prediction(skills_list):
    col_list = joblib.load(DATA_COL)
    pred_df = pd.DataFrame(col_list.isin(skills_list)).T
    pred_df = pred_df.astype(int)
    pred_df.columns = col_list
    pred_df['experanse_years'] = 1

    pred_df_scl = StandardScaler().fit_transform(pred_df.T)

    model = joblib.load(final_model_path)

    result = model.predict(pred_df_scl.T)

    return result[0]

def prediction_prob(skills_list):
    col_list = joblib.load(DATA_COL)
    pred_df = pd.DataFrame(col_list.isin(skills_list)).T
    pred_df = pred_df.astype(int)
    pred_df.columns = X.columns
    pred_df['experanse_years'] = 1

    pred_df_scl = StandardScaler().fit_transform(pred_df.T)

    model = joblib.load(final_model_path)

    class_names = joblib.load(Classes_NAME)
    probabilities = model.predict_proba(pred_df_scl.T)
    class_prob_dict = dict(zip(class_names, probabilities.flatten()))
    # Create a DataFrame from the dictionary
    df = pd.DataFrame(list(class_prob_dict.items()),
                      columns=['Class', 'Probability'])

    # Sort the DataFrame by the 'Probability' column in ascending order
    df_sorted = df.sort_values(by='Probability', ascending=False)

    return df_sorted

prediction_prob(skills_list)

# get skills lists


df

unique_values = {}
for col in list_columns:
    unique = df[col].explode().unique()
    unique_values[col] = list(unique)

unique_values.keys()

# list(df['EdLevel'].unique())

# key_Edlevel = list(df['EdLevel'].unique())
Key_languages = unique_values['Key_languages']
Key_Databases = unique_values['Databases']
key_Platforms = unique_values['Platforms']
key_Web_frames = unique_values['Web_frames']
key_varied_Tech = unique_values['varied_Tech']
key_Tools = unique_values['Tools']
key_collabTools = unique_values['collabTools']

# joblib.dump(key_collabTools, COLLABTOOLS_PATH)

# joblib.load(LANGUAGES_PATH)

skills_clusters = joblib.load(skills_clusters_path)

# Make find_matching_skill fun


skills_list

test = ['R', 'Python', 'TensorFlow', 'Jupyter Notebook/JupyterLab']

set(test)

skills_clusters.to_dict()

def find_matching_skill_groups(skill_list, skill_dict):
    matching_skill_groups = []

    for group_name, group_skills in skill_dict.items():
        common_skills = set(skill_list) & set(group_skills)
        if len(common_skills) >= 3:
            non_common_skills = list(set(group_skills) - set(common_skills))
            matching_skill_groups.append((group_name, non_common_skills))

    return matching_skill_groups


# Your list of skills
skills_list = ['R', 'Python', 'TensorFlow',
               'Jupyter Notebook/JupyterLab', 'pandas', 'Numpy']

# Your dictionary of skill groups
skills_dict = skills_clusters

# Find matching skill groups
matching_groups = find_matching_skill_groups(skills_list, skills_dict)

# Display matching skill groups and list without common skills
if matching_groups:
    print("Matching skill groups:")
    for group_name, non_common_skills in matching_groups:
        print(f"Group: {group_name}")
        print(f"Original Skills: {skills_dict[group_name]}")
        print(f"Non-Common Skills: {non_common_skills}")
        print("---")
else:
    print("No matching skill groups found.")

---


# Now lets move to predict model pipline --> src\models\predict_model.py
